Tic Tac Toe game mode
This game mode is turn-based, switching between player 1 and player 2. On their
turn, each player kicks a ball through a square. If the square is unclaimed, it
will light up with that player's color. If one of the players claims 3 squares
in a row (vertically, horizontally, or diagonally), they win. The winning row
of sqaures flashes off and on a few times to indicate the win.